[MaL] ODE: Move out Handles implementation.
